Button Code
Here is the code for my button. Not all of it is working, for some reason the use of html within the components file creates problems. Outside of that folder it doesn't give me the same errors, I haven't been able to figure that one out yet but just wanted to make sure I got what I have into you.